added ticket requirements
Added game logic to ensure that the join / host buttons only pop up when specific requirements are met.

case 1 - player name is not empty.
- if not empty, grant access to the host button.
case 2 - player name goes from filled in to empty.
- host button & join button goes from visible to invisible.
case 3 - a room code is entered with a name.
- host button shows up
case 4 - a room code is entered with a name before being removed from the field
- host button disappears
case 5 - both entries are filled in but are then empty
- both join and host buttons with go from visible to invisible.