Join logic
Added so that as soon as a user enters in the room code, the only provided option is to join. As soon as it is empty, it would offer the host option with the provided option as "Join" If both fields are empty, no option for join/host will be provided.